requirements.txt
pandas
nltk
sentence-transformers
transformers
torch
pinecone-client
fastapi
uvicorn
streamlit
langgraph
python-dotenv
